4h 50m - Advanced Enemy Spawner Logic
Added:

- Enemy Spawn Blocks. Spawn blocks define enemy spawn behaviors' over a given time period.
- Enemy spawn rate curve
- Added normal enemy spawn type weighted probability.
- Scaffolding for elite spawns.
- Scaffolding for boss spawns.
- Win game state

Changes:

- Enemy managers now uses EnemySpawnBlocks instead of spawning enemies directly.

Known Issues:

- Spawn Blocks can not be 1 or less seconds in duration.

Notes:

30 minutes was wasted trying to implement a more complicated version of the enemy spawner, but it required SerializeReference and Polymorphism, which is outside the scope of the Prototype.

About 20 minutes was used chasing an index out of range error. Most likely due to the use of coroutines for spawning enemies. The Update loops of the manager classes are getting bloated, and inproper logic in the if statements are starting to collide with one another.