40h 00m - Added Progress Bar UI.
Added:

- Added progress bar UI that shows how close you are to completing a block.
- Balanced store items.

Notes:

This took 3 hours because I also tried to add wave markers as well, however this lead to quite complicated scaling for the progression bar which ultimately caused too many bugs and had to be removed. This would have been quite a bit easier if we had events, as we're manually polling for everything instead of just subscribing to WaveCompleted() EnemyKilled() etc.